backend/requirements.txt
annotated-types==0.7.0
faiss-cpu>=1.7.4
pillow-heif>=0.13.0
python-pptx>=0.6.23
watchdog>=4.0.0
anyio==4.9.0
attrs==18.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
distro==1.9.0
filelock==3.18.0
fpdf==1.7.2
fsspec==2025.5.0
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==1.79.0
packaging==25.0
pandoc==2.4
pdflatex==0.1.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pluggy==1.6.0
plumbum==1.9.0
ply==3.11
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.3.5
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0